requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
sqlalchemy==2.0.36
pypdf==5.1.0
python-multipart==0.0.17
sentence-transformers==3.3.1
numpy==2.1.3
faiss-cpu==1.13.2
openai==1.57.4
python-dotenv==1.0.1